fix: challengeId를 Integer로 받았던 오류 수정

+++ backend/src/main/java/km/cd/backend/challenge/repository/ChallengeRepository.java
@@ -3,6 +3,6 @@ package km.cd.backend.challenge.repository;
 import km.cd.backend.challenge.domain.Challenge;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ChallengeRepository extends CrudRepository<Challenge, Integer> {
+public interface ChallengeRepository extends CrudRepository<Challenge, Long> {
 
 }
